pyproject.toml
[project]
name = "clawdia"
version = "0.1.0"
description = "Open-source Raspberry Pi voice assistant with IR TV control"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pydantic-ai-slim[openrouter]>=1.0",
    "openai>=1.0",
    "python-telegram-bot>=22.0",
    "httpx>=0.27",
    "spotipy>=2.24",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "loguru>=0.7",
]

[project.optional-dependencies]
voice = [
    "openwakeword>=0.6",
    "pyaudio>=0.2",
    "numpy>=1.24",
    "scipy>=1.10",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.0",
    "ruff>=0.8",
]

[project.scripts]
clawdia = "clawdia.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawdia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
